requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers
Pillow>=10.2.0
django-allauth
dj-rest-auth
requests
